requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
awscli==1.35.17
aylak==0.0.6
boto==2.49.0
botocore==1.35.51
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
distro==1.9.0
docutils==0.16
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.50.4
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.52.2
packaging==24.1
pillow==10.2.0
playwright==1.48.0
pm2==0.0.4.4
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonansi==1.0.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
rsa==4.7.2
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
Telethon==1.37.0
tiktoken==0.8.0
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0
zipp==3.20.2
